requirements.txt
streamlit
gspread
google-auth
google-api-python-client 
google-auth-oauthlib